feat: Выполнена задача, добавлены новые экраны в игре
