--- packages/client/src/pages/End/End.tsx
@@ -1,11 +1,0 @@
-import { PageLayout } from '@/shared/layouts'
-
-const EndPage = () => {
-  return (
-    <PageLayout>
-      <h1>End Page</h1>
-    </PageLayout>
-  )
-}
-
-export default EndPage
