--- packages/client/src/pages/End/index.tsx
@@ -1,1 +1,0 @@
-export { default as EndPage } from './End'
