+++ packages/client/src/pages/Game/components/GameEngine/utils/generateEmptyBoard.ts
@@ -1,4 +1,4 @@
-import { GameBoard } from '../models.js'
+import { GameBoard } from '../models.ts'
 
 export const generateEmptyBoard = (size: number): GameBoard => {
   return new Array(size).fill(null).map(() => new Array(size).fill(0))
