+++ packages/client/src/pages/Game/components/GameEngine/utils/moveLeft.ts
@@ -1,4 +1,4 @@
-import { GameBoard } from '../models.js'
+import { GameBoard } from '../models.ts'
 
 export const moveLeft = (
   size: number,
