+++ packages/client/src/pages/Game/components/GameEngine/utils/rotateBoard.ts
@@ -1,5 +1,5 @@
-import { GameBoard } from '../models.js'
-import { generateEmptyBoard } from './generateEmptyBoard.js'
+import { GameBoard } from '../models.ts'
+import { generateEmptyBoard } from './generateEmptyBoard.ts'
 
 export const rotateBoard = (size: number, board: GameBoard): GameBoard => {
   const newBoard = generateEmptyBoard(size)
