+++ packages/client/src/pages/Game/index.ts
@@ -1,0 +1,1 @@
+export { default as GamePage } from './Game'
