--- packages/client/src/pages/Game/index.tsx
@@ -1,1 +1,0 @@
-export { default as GamePage } from './Game.js'
