+++ packages/client/src/pages/Game/style.module.css
@@ -1,0 +1,73 @@
+.gameBackground {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  min-height: 100vh;
+  padding: 50px 0;
+  overflow: scroll;
+  background: url('/assets/background.png') center center / cover;
+  width: 100vw;
+}
+.outButton {
+  position: absolute;
+  left: 30px;
+  top: 30px;
+  background: black;
+  border-radius: 50%;
+  padding: 12px;
+}
+.countdown {
+  padding: 30px 60px;
+  min-width: 157px;
+  font-size: 64px;
+  color: #333;
+  backdrop-filter: blur(10px);
+  -webkit-backdrop-filter: blur(10px);
+  overflow: hidden;
+  border-radius: 20px;
+  box-shadow: inset -0.75px -0.5px rgba(255, 255, 255, 0.1),
+    inset + 0.75px+0.5px rgba(255, 255, 255, 0.025),
+    3px 2px 10px rgba(0, 0, 0, 0.25);
+}
+
+.content {
+  padding: 30px;
+  margin-top: 5vh;
+  backdrop-filter: blur(10px);
+  -webkit-backdrop-filter: blur(10px);
+  overflow: auto;
+  border-radius: 20px;
+  box-shadow: inset -0.75px -0.5px rgba(255, 255, 255, 0.1),
+    inset + 0.75px+0.5px rgba(255, 255, 255, 0.025),
+    3px 2px 10px rgba(0, 0, 0, 0.25);
+  text-align: center;
+  max-height: 90vh;
+  transition: opacity 0.5s ease, transform 0.5s ease;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+.content h1 {
+  font-size: 2em;
+}
+.content h2 {
+  font-size: 1.4em;
+}
+
+.instructions {
+  margin-top: 20px;
+  max-width: 700px;
+  line-height: 1.5;
+  text-align: left;
+  background: rgba(255, 255, 255, 0.94);
+  padding: 20px;
+  border-radius: 10px;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+}
+.instructions ul {
+  list-style: disc;
+}
+.instructions li {
+  margin-left: 27px;
+}
