+++ packages/client/src/pages/Game/types.ts
@@ -1,0 +1,12 @@
+export enum GAME_STATE {
+  START = 'START',
+  IN_GAME = 'IN_GAME',
+  ENDGAME = 'ENDGAME',
+}
+
+export enum GameStatus {
+  idle = 'idle',
+  playing = 'playing',
+  win = 'win',
+  lose = 'lose',
+}
