+++ packages/client/src/shared/ui/Button/tabs.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react'
 import * as TabsPrimitive from '@radix-ui/react-tabs'
-
-import { cn } from '@/lib/utils'
+import { cn } from '@/shared/lib/utils'
 
 function Tabs({
   className,
